Code validate and folder name
