+++ roulette/src/main/java/com/masiv/cleancode/roulette/Utils/Result.java
@@ -1,0 +1,6 @@
+package com.masiv.cleancode.roulette.Utils;
+
+public enum Result {
+    WIN,
+    LOSE
+}
