--- roulette/src/main/java/com/masiv/cleancode/roulette/utils/Result.java
@@ -1,6 +1,0 @@
-package com.masiv.cleancode.roulette.utils;
-
-public enum Result {
-    WIN,
-    LOSE
-}
